# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SaaS API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface RegisterBody {
  email: string;
  /** @minLength 6 */
  password: string;
  name: string;
}

export interface LoginBody {
  email: string;
  password: string;
}

export interface User {
  id: number;
  email: string;
  name: string;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface Client {
  id: number;
  userId: number;
  name: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateClientBody {
  name: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  notes?: string | null;
}

export interface UpdateClientBody {
  name?: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  notes?: string | null;
}

export type SessionStatus = (typeof SessionStatus)[keyof typeof SessionStatus];

export const SessionStatus = {
  pending: "pending",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export interface Session {
  id: number;
  userId: number;
  clientId: number;
  /** @nullable */
  clientName?: string | null;
  date: string;
  status: SessionStatus;
  price: number;
  paid: boolean;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
  updatedAt: string;
}

export type CreateSessionBodyStatus =
  (typeof CreateSessionBodyStatus)[keyof typeof CreateSessionBodyStatus];

export const CreateSessionBodyStatus = {
  pending: "pending",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export interface CreateSessionBody {
  clientId: number;
  date: string;
  status?: CreateSessionBodyStatus;
  price: number;
  paid?: boolean;
  /** @nullable */
  notes?: string | null;
}

export type UpdateSessionBodyStatus =
  (typeof UpdateSessionBodyStatus)[keyof typeof UpdateSessionBodyStatus];

export const UpdateSessionBodyStatus = {
  pending: "pending",
  completed: "completed",
  cancelled: "cancelled",
} as const;

export interface UpdateSessionBody {
  clientId?: number;
  date?: string;
  status?: UpdateSessionBodyStatus;
  price?: number;
  paid?: boolean;
  /** @nullable */
  notes?: string | null;
}

export interface DashboardStats {
  totalClients: number;
  totalSessions: number;
  weeklySessions: number;
  monthlyRevenue: number;
  unpaidRevenue: number;
}

export interface MonthlyRevenue {
  month: string;
  revenue: number;
  sessions: number;
}

export type ListSessionsParams = {
  /**
   * @nullable
   */
  clientId?: number | null;
};