# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SaaS API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication operations
  - name: clients
    description: Client management
  - name: sessions
    description: Session management
  - name: dashboard
    description: Dashboard statistics
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: Registered successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: User already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /clients:
    get:
      operationId: listClients
      tags: [clients]
      summary: List all clients
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of clients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Client"
    post:
      operationId: createClient
      tags: [clients]
      summary: Create a client
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClientBody"
      responses:
        "201":
          description: Client created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /clients/{id}:
    get:
      operationId: getClient
      tags: [clients]
      summary: Get a client
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
        "404":
          description: Client not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateClient
      tags: [clients]
      summary: Update a client
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateClientBody"
      responses:
        "200":
          description: Updated client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
        "404":
          description: Client not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteClient
      tags: [clients]
      summary: Delete a client
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Client not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /sessions:
    get:
      operationId: listSessions
      tags: [sessions]
      summary: List sessions
      security:
        - bearerAuth: []
      parameters:
        - name: clientId
          in: query
          required: false
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: List of sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Session"
    post:
      operationId: createSession
      tags: [sessions]
      summary: Create a session
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSessionBody"
      responses:
        "201":
          description: Session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /sessions/{id}:
    get:
      operationId: getSession
      tags: [sessions]
      summary: Get a session
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateSession
      tags: [sessions]
      summary: Update a session
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSessionBody"
      responses:
        "200":
          description: Updated session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteSession
      tags: [sessions]
      summary: Delete a session
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /dashboard/stats:
    get:
      operationId: getDashboardStats
      tags: [dashboard]
      summary: Get dashboard statistics
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Dashboard stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"

  /dashboard/monthly-revenue:
    get:
      operationId: getMonthlyRevenue
      tags: [dashboard]
      summary: Get monthly revenue for last 6 months
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Monthly revenue data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MonthlyRevenue"

  /dashboard/recent-sessions:
    get:
      operationId: getRecentSessions
      tags: [dashboard]
      summary: Get recent sessions
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Recent sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Session"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    RegisterBody:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        name:
          type: string
      required:
        - email
        - password
        - name

    LoginBody:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required:
        - email
        - password

    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - email
        - name
        - createdAt

    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required:
        - token
        - user

    Client:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        phone:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - name
        - createdAt
        - updatedAt

    CreateClientBody:
      type: object
      properties:
        name:
          type: string
        phone:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
      required:
        - name

    UpdateClientBody:
      type: object
      properties:
        name:
          type: string
        phone:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]

    Session:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        clientId:
          type: integer
        clientName:
          type: ["string", "null"]
        date:
          type: string
          format: date-time
        status:
          type: string
          enum: [pending, completed, cancelled]
        price:
          type: number
        paid:
          type: boolean
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - clientId
        - date
        - status
        - price
        - paid
        - createdAt
        - updatedAt

    CreateSessionBody:
      type: object
      properties:
        clientId:
          type: integer
        date:
          type: string
          format: date-time
        status:
          type: string
          enum: [pending, completed, cancelled]
        price:
          type: number
        paid:
          type: boolean
        notes:
          type: ["string", "null"]
      required:
        - clientId
        - date
        - price

    UpdateSessionBody:
      type: object
      properties:
        clientId:
          type: integer
        date:
          type: string
          format: date-time
        status:
          type: string
          enum: [pending, completed, cancelled]
        price:
          type: number
        paid:
          type: boolean
        notes:
          type: ["string", "null"]

    DashboardStats:
      type: object
      properties:
        totalClients:
          type: integer
        totalSessions:
          type: integer
        weeklySessions:
          type: integer
        monthlyRevenue:
          type: number
        unpaidRevenue:
          type: number
      required:
        - totalClients
        - totalSessions
        - weeklySessions
        - monthlyRevenue
        - unpaidRevenue

    MonthlyRevenue:
      type: object
      properties:
        month:
          type: string
        revenue:
          type: number
        sessions:
          type: integer
      required:
        - month
        - revenue
        - sessions