# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SaaS API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Register a new user
 */
export const registerBodyPasswordMin = 6;

export const RegisterBody = zod.object({
  email: zod.string().email(),
  password: zod.string().min(registerBodyPasswordMin),
  name: zod.string(),
});

/**
 * @summary Login
 */
export const LoginBody = zod.object({
  email: zod.string().email(),
  password: zod.string(),
});

export const LoginResponse = zod.object({
  token: zod.string(),
  user: zod.object({
    id: zod.number(),
    email: zod.string(),
    name: zod.string(),
    createdAt: zod.coerce.date(),
  }),
});

/**
 * @summary Get current user
 */
export const GetMeResponse = zod.object({
  id: zod.number(),
  email: zod.string(),
  name: zod.string(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary List all clients
 */
export const ListClientsResponseItem = zod.object({
  id: zod.number(),
  userId: zod.number(),
  name: zod.string(),
  phone: zod.string().nullish(),
  notes: zod.string().nullish(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});
export const ListClientsResponse = zod.array(ListClientsResponseItem);

/**
 * @summary Create a client
 */
export const CreateClientBody = zod.object({
  name: zod.string(),
  phone: zod.string().nullish(),
  notes: zod.string().nullish(),
});

/**
 * @summary Get a client
 */
export const GetClientParams = zod.object({
  id: zod.coerce.number(),
});

export const GetClientResponse = zod.object({
  id: zod.number(),
  userId: zod.number(),
  name: zod.string(),
  phone: zod.string().nullish(),
  notes: zod.string().nullish(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});

/**
 * @summary Update a client
 */
export const UpdateClientParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateClientBody = zod.object({
  name: zod.string().optional(),
  phone: zod.string().nullish(),
  notes: zod.string().nullish(),
});

export const UpdateClientResponse = zod.object({
  id: zod.number(),
  userId: zod.number(),
  name: zod.string(),
  phone: zod.string().nullish(),
  notes: zod.string().nullish(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});

/**
 * @summary Delete a client
 */
export const DeleteClientParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary List sessions
 */
export const ListSessionsQueryParams = zod.object({
  clientId: zod.coerce.number().nullish(),
});

export const ListSessionsResponseItem = zod.object({
  id: zod.number(),
  userId: zod.number(),
  clientId: zod.number(),
  clientName: zod.string().nullish(),
  date: zod.coerce.date(),
  status: zod.enum(["pending", "completed", "cancelled"]),
  price: zod.number(),
  paid: zod.boolean(),
  notes: zod.string().nullish(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});
export const ListSessionsResponse = zod.array(ListSessionsResponseItem);

/**
 * @summary Create a session
 */
export const CreateSessionBody = zod.object({
  clientId: zod.number(),
  date: zod.coerce.date(),
  status: zod.enum(["pending", "completed", "cancelled"]).optional(),
  price: zod.number(),
  paid: zod.boolean().optional(),
  notes: zod.string().nullish(),
});

/**
 * @summary Get a session
 */
export const GetSessionParams = zod.object({
  id: zod.coerce.number(),
});

export const GetSessionResponse = zod.object({
  id: zod.number(),
  userId: zod.number(),
  clientId: zod.number(),
  clientName: zod.string().nullish(),
  date: zod.coerce.date(),
  status: zod.enum(["pending", "completed", "cancelled"]),
  price: zod.number(),
  paid: zod.boolean(),
  notes: zod.string().nullish(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});

/**
 * @summary Update a session
 */
export const UpdateSessionParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateSessionBody = zod.object({
  clientId: zod.number().optional(),
  date: zod.coerce.date().optional(),
  status: zod.enum(["pending", "completed", "cancelled"]).optional(),
  price: zod.number().optional(),
  paid: zod.boolean().optional(),
  notes: zod.string().nullish(),
});

export const UpdateSessionResponse = zod.object({
  id: zod.number(),
  userId: zod.number(),
  clientId: zod.number(),
  clientName: zod.string().nullish(),
  date: zod.coerce.date(),
  status: zod.enum(["pending", "completed", "cancelled"]),
  price: zod.number(),
  paid: zod.boolean(),
  notes: zod.string().nullish(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});

/**
 * @summary Delete a session
 */
export const DeleteSessionParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Get dashboard statistics
 */
export const GetDashboardStatsResponse = zod.object({
  totalClients: zod.number(),
  totalSessions: zod.number(),
  weeklySessions: zod.number(),
  monthlyRevenue: zod.number(),
  unpaidRevenue: zod.number(),
});

/**
 * @summary Get monthly revenue for last 6 months
 */
export const GetMonthlyRevenueResponseItem = zod.object({
  month: zod.string(),
  revenue: zod.number(),
  sessions: zod.number(),
});
export const GetMonthlyRevenueResponse = zod.array(
  GetMonthlyRevenueResponseItem,
);

/**
 * @summary Get recent sessions
 */
export const GetRecentSessionsResponseItem = zod.object({
  id: zod.number(),
  userId: zod.number(),
  clientId: zod.number(),
  clientName: zod.string().nullish(),
  date: zod.coerce.date(),
  status: zod.enum(["pending", "completed", "cancelled"]),
  price: zod.number(),
  paid: zod.boolean(),
  notes: zod.string().nullish(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});
export const GetRecentSessionsResponse = zod.array(
  GetRecentSessionsResponseItem,
);